+++ src/levels.cpp
@@ -952,7 +952,7 @@ bool levLoadData(char const *name, Sha256 const *hash, char *pSaveName, GAME_TYP
 				debug(LOG_WZ, "Init mission stuff");
 				if (!startMissionSave(psNewLevel->type))
 				{
-					debug(LOG_ERROR, "Failed startMissionSave(%d)!", psNewLevel->type);
+					debug(LOG_ERROR, "Failed startMissionSave(%d)!", static_cast<int8_t>(psNewLevel->type));
 					return false;
 				}
 
@@ -973,7 +973,7 @@ bool levLoadData(char const *name, Sha256 const *hash, char *pSaveName, GAME_TYP
 			debug(LOG_NEVER, "Start mission - no .gam");
 			if (!startMission((LEVEL_TYPE)psNewLevel->type, nullptr))
 			{
-				debug(LOG_ERROR, "Failed startMission(%d)!", psNewLevel->type);
+				debug(LOG_ERROR, "Failed startMission(%d)!", static_cast<int8_t>(psNewLevel->type));
 				return false;
 			}
 		}
@@ -1030,7 +1030,7 @@ bool levLoadData(char const *name, Sha256 const *hash, char *pSaveName, GAME_TYP
 					debug(LOG_WZ, "Init mission stuff");
 					if (!startMissionSave(psNewLevel->type))
 					{
-						debug(LOG_ERROR, "Failed startMissionSave(%d)!", psNewLevel->type);
+						debug(LOG_ERROR, "Failed startMissionSave(%d)!", static_cast<uint8_t>(psNewLevel->type));
 						return false;
 					}
 
@@ -1057,7 +1057,7 @@ bool levLoadData(char const *name, Sha256 const *hash, char *pSaveName, GAME_TYP
 					debug(LOG_WZ, "LDS_COMPLETE / LDS_CAMSTART");
 					if (!startMission(LEVEL_TYPE::LDS_CAMSTART, psNewLevel->apDataFiles[i]))
 					{
-						debug(LOG_ERROR, "Failed startMission(%d, %s)!", LEVEL_TYPE::LDS_CAMSTART, psNewLevel->apDataFiles[i]);
+						debug(LOG_ERROR, "Failed startMission(%d, %s)!", static_cast<int8_t>(LEVEL_TYPE::LDS_CAMSTART), psNewLevel->apDataFiles[i]);
 						return false;
 					}
 					break;
@@ -1065,7 +1065,7 @@ bool levLoadData(char const *name, Sha256 const *hash, char *pSaveName, GAME_TYP
 					debug(LOG_WZ, "LDS_BETWEEN");
 					if (!startMission(LEVEL_TYPE::LDS_BETWEEN, psNewLevel->apDataFiles[i]))
 					{
-						debug(LOG_ERROR, "Failed startMission(%d, %s)!", LEVEL_TYPE::LDS_BETWEEN, psNewLevel->apDataFiles[i]);
+						debug(LOG_ERROR, "Failed startMission(%d, %s)!", static_cast<int8_t>(LEVEL_TYPE::LDS_BETWEEN), psNewLevel->apDataFiles[i]);
 						return false;
 					}
 					break;
@@ -1074,7 +1074,7 @@ bool levLoadData(char const *name, Sha256 const *hash, char *pSaveName, GAME_TYP
 					debug(LOG_WZ, "LDS_MKEEP");
 					if (!startMission(LEVEL_TYPE::LDS_MKEEP, psNewLevel->apDataFiles[i]))
 					{
-						debug(LOG_ERROR, "Failed startMission(%d, %s)!", LEVEL_TYPE::LDS_MKEEP, psNewLevel->apDataFiles[i]);
+						debug(LOG_ERROR, "Failed startMission(%d, %s)!", static_cast<int8_t>(LEVEL_TYPE::LDS_MKEEP), psNewLevel->apDataFiles[i]);
 						return false;
 					}
 					break;
@@ -1082,7 +1082,7 @@ bool levLoadData(char const *name, Sha256 const *hash, char *pSaveName, GAME_TYP
 					debug(LOG_WZ, "LDS_CAMCHANGE");
 					if (!startMission(LEVEL_TYPE::LDS_CAMCHANGE, psNewLevel->apDataFiles[i]))
 					{
-						debug(LOG_ERROR, "Failed startMission(%d, %s)!", LEVEL_TYPE::LDS_CAMCHANGE, psNewLevel->apDataFiles[i]);
+						debug(LOG_ERROR, "Failed startMission(%d, %s)!", static_cast<int8_t>(LEVEL_TYPE::LDS_CAMCHANGE), psNewLevel->apDataFiles[i]);
 						return false;
 					}
 					break;
@@ -1091,7 +1091,7 @@ bool levLoadData(char const *name, Sha256 const *hash, char *pSaveName, GAME_TYP
 					debug(LOG_WZ, "LDS_EXPAND");
 					if (!startMission(LEVEL_TYPE::LDS_EXPAND, psNewLevel->apDataFiles[i]))
 					{
-						debug(LOG_ERROR, "Failed startMission(%d, %s)!", LEVEL_TYPE::LDS_EXPAND, psNewLevel->apDataFiles[i]);
+						debug(LOG_ERROR, "Failed startMission(%d, %s)!", static_cast<int8_t>(LEVEL_TYPE::LDS_EXPAND), psNewLevel->apDataFiles[i]);
 						return false;
 					}
 					break;
@@ -1099,7 +1099,7 @@ bool levLoadData(char const *name, Sha256 const *hash, char *pSaveName, GAME_TYP
 					debug(LOG_WZ, "LDS_LIMBO");
 					if (!startMission(LEVEL_TYPE::LDS_EXPAND_LIMBO, psNewLevel->apDataFiles[i]))
 					{
-						debug(LOG_ERROR, "Failed startMission(%d, %s)!", LEVEL_TYPE::LDS_EXPAND_LIMBO, psNewLevel->apDataFiles[i]);
+						debug(LOG_ERROR, "Failed startMission(%d, %s)!", static_cast<int8_t>(LEVEL_TYPE::LDS_EXPAND_LIMBO), psNewLevel->apDataFiles[i]);
 						return false;
 					}
 					break;
@@ -1108,7 +1108,7 @@ bool levLoadData(char const *name, Sha256 const *hash, char *pSaveName, GAME_TYP
 					debug(LOG_WZ, "LDS_MCLEAR");
 					if (!startMission(LEVEL_TYPE::LDS_MCLEAR, psNewLevel->apDataFiles[i]))
 					{
-						debug(LOG_ERROR, "Failed startMission(%d, %s)!", LEVEL_TYPE::LDS_MCLEAR, psNewLevel->apDataFiles[i]);
+						debug(LOG_ERROR, "Failed startMission(%d, %s)!", static_cast<int8_t>(LEVEL_TYPE::LDS_MCLEAR), psNewLevel->apDataFiles[i]);
 						return false;
 					}
 					break;
@@ -1116,7 +1116,7 @@ bool levLoadData(char const *name, Sha256 const *hash, char *pSaveName, GAME_TYP
 					debug(LOG_WZ, "LDS_MKEEP_LIMBO");
 					if (!startMission(LEVEL_TYPE::LDS_MKEEP_LIMBO, psNewLevel->apDataFiles[i]))
 					{
-						debug(LOG_ERROR, "Failed startMission(%d, %s)!", LEVEL_TYPE::LDS_MKEEP_LIMBO, psNewLevel->apDataFiles[i]);
+						debug(LOG_ERROR, "Failed startMission(%d, %s)!", static_cast<int8_t>(LEVEL_TYPE::LDS_MKEEP_LIMBO), psNewLevel->apDataFiles[i]);
 						return false;
 					}
 					break;
@@ -1125,7 +1125,7 @@ bool levLoadData(char const *name, Sha256 const *hash, char *pSaveName, GAME_TYP
 					debug(LOG_WZ, "default (MULTIPLAYER)");
 					if (!startMission(LEVEL_TYPE::LDS_CAMSTART, psNewLevel->apDataFiles[i]))
 					{
-						debug(LOG_ERROR, "Failed startMission(%d, %s) (default)!", LEVEL_TYPE::LDS_CAMSTART, psNewLevel->apDataFiles[i]);
+						debug(LOG_ERROR, "Failed startMission(%d, %s) (default)!", static_cast<int8_t>(LEVEL_TYPE::LDS_CAMSTART), psNewLevel->apDataFiles[i]);
 						return false;
 					}
 					break;
@@ -1155,7 +1155,7 @@ bool levLoadData(char const *name, Sha256 const *hash, char *pSaveName, GAME_TYP
 		//load MidMission Extras
 		if (!loadMissionExtras(pSaveName, psNewLevel->type))
 		{
-			debug(LOG_ERROR, "Failed loadMissionExtras(%s, %d)!", pSaveName, psNewLevel->type);
+			debug(LOG_ERROR, "Failed loadMissionExtras(%s, %d)!", pSaveName, static_cast<int8_t>(psNewLevel->type));
 			return false;
 		}
 	}
